Improvement Suggestion
Hi team, I would like to suggest some changes related to disable-virtualization documentation like explaining the impact of setting true or false the virtualization and also related with "position: relative" style hardcoded which will be solved if you have a container inside. This was something that we faced when building an autocomplete. Please talk with me via slack @miguel.vicente